fix: preserve CU metadata until finalization, fix latencyMs type in IPCMessages (#6912)

+++ clients/shared/IPC/IPCMessages.swift
@@ -762,7 +762,7 @@ extension IPCMemoryRecalled {
         selectedCount: Int,
         rerankApplied: Bool,
         injectedTokens: Int,
-        latencyMs: Double,
+        latencyMs: Int,
         topCandidates: [IPCMemoryRecalledCandidateDebug]
     ) {
         self.init(
